build-requirements.txt
autopep8
coverage
mock
path.py
pip
pycodestyle
pyflakes
pytest
pytest-cov
pytest-mock
python-coveralls
setuptools